+++ src/main/java/baseball/domain/number/NumberConstant.java
@@ -1,0 +1,7 @@
+package baseball.domain.number;
+
+public class NumberConstant {
+    public static final Integer MAX_NUMBER = 9;
+    public static final Integer MIN_NUMBER = 1;
+    public static final Integer NUMBER_COUNT = 3;
+}
